refactor: move readFiles to a separate file
