+++ packages/@vue/cli/lib/upgrade.js
@@ -197,6 +197,10 @@ async function upgradeAll (context) {
   }
 
   // TODO: upgrade all (interactive)
+  // for patch & minor versions, upgrade directly
+  // for major versions, prompt before upgrading
+
+  // TODO: format the output https://github.com/angular/angular-cli/blob/34a55c96b2ed38b226879913839b97c601387653/packages/schematics/update/update/index.ts#L490-L509
   log(upgradable)
 
   failSpinner()
